Przeorganizowano projekt. Usuniêto niektóre pliki.
