--- src/main/java/edu/kis/vh/stacks/factory/IStacksFactory.java
@@ -1,17 +1,0 @@
-package edu.kis.vh.stacks.factory;
-
-import edu.kis.vh.stacks.Stack;
-import edu.kis.vh.stacks.StackFIFO;
-import edu.kis.vh.stacks.StackHanoi;
-
-public interface IStacksFactory {
-
-	public Stack getStandardStack();
-
-	public Stack getFalseStack();
-
-	public StackFIFO getFIFOStack();
-
-	public StackHanoi getHanoiStack();
-
-}
